pyproject.toml
[tool.poetry]
name = "ansible-my-wsl"
description = "Ansible WSL workstation"
version = "1.0.0"
license = "MIT"

authors = ["Sergei Aleksandrov <sergei.a.aleks@gmail.com>"]

readme = "README.md"
keywords = [
  "ansible",
  "wsl",
  "workstation",
  "playbooks",
  "molecule",
  "linux",
  "debian"
]

homepage = "https://github.com/sjlex/ansible-my-wsl"
repository = "https://github.com/sjlex/ansible-my-wsl"
documentation = "https://github.com/sjlex/ansible-my-wsl/blob/main/README.md"

[tool.poetry.dependencies]
ansible = "8.0.0"

[tool.poetry.group.dev.dependencies]
python = "3.11.4"
molecule = "5.0.1"
molecule-plugins = { version = "23.4.1", extras = ["docker", "vagrant"] }
pytest-testinfra = "8.1.0"
ansible-lint = { version = "6.16.2", markers = "platform_system != 'Windows'" }
yamllint = "1.32.0"
ruff = "0.0.274"
black = "23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
